Fix RPM lint problem.
The timestamp embedded in the python bytecode file (mtime of
original source in 2006) isn't equal to the mtime of the installed
source file (now), which will force the interpreter to recompile
the .py source every time, ignoring the saved bytecode.

This patch should fix it by moving compilation to the install
phase.